chore: move usePayPalVisibility composable into its own folder
