+++ apps/web/modules/paypal/runtime/composables/usePayPal/index.ts
@@ -1,2 +1,1 @@
 export * from './usePayPal';
-export * from './usePayPalVisibility';
