+++ apps/web/modules/paypal/runtime/composables/usePayPalVisibility/index.ts
@@ -1,0 +1,1 @@
+export * from './usePayPalVisibility';
